s3_redshift_lambda/requirements.txt
python-dotenv==0.19.0
psycopg2-binary==2.8.3